moved animation component
